% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domestic_ranking.R
\name{domesticRank}
\alias{domesticRank}
\title{Scraps mojoboxoffice.com for revenue data}
\usage{
domesticRank(pages, ruta = "~/Domestic_rank.csv")
}
\arguments{
\item{pages}{numeric. Number of pages you want to scrap from box office mojo's domestic ranking.}

\item{ruta}{character. A valid path in your computer where you want to create the csv file. By default the csv named mojo will be created in the working directory.}
}
\value{
It returns a csv in the specified path
}
\description{
This function scraps boxofficemojo.com and downloads rank, lifetime gross box office data, year, and id from the domestic ranking: 'http://www.boxofficemojo.com/alltime/domestic.htm'.
}
