% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/international.R
\name{international}
\alias{international}
\title{Scraps mojoboxoffice.com for revenue data}
\usage{
international(ctry, wk, yr, path, serverHelp = TRUE)
}
\arguments{
\item{ctry}{character. Country from which you want to scrap revenue data. Make sure you type it the exact same way as the url. For example: http://www.boxofficemojo.com/intl/australia/?yr=2018&wk=23&p=.htm}

\item{wk}{integer. A numeric value or vector that indicates which weeks of the year you want to scrap. For example: 34 will scrap the 34th week. 1:52 will scrap the 52 weeks of the given year.}

\item{yr}{integer.A numeric value or vector that indicates which year you want to scrap.}

\item{path}{character. A valid path in your computer where you want to create the csv file.}

\item{serverHelp}{logical. Indicates if you want mojoscRaper to initiate a Docker server or you will initiate it youself. TRUE by default. Remember that this function NEEDS a Docker server running}
}
\value{
It returns a csv in the specified path
}
\description{
This function scraps boxofficemojo.com and downloads rank, lifetime gross box office data, year, and id from the domestic ranking: 'http://www.boxofficemojo.com/alltime/domestic.htm'.
}
