% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mojo.R
\name{mojo}
\alias{mojo}
\title{Scraps mojoboxoffice.com for revenue data}
\usage{
mojo(from, to, ruta = "~/mojo.csv")
}
\arguments{
\item{from}{character. A date in YYYY-MM-DD format.}

\item{to}{character. A date in YYYY-MM-DD format.}

\item{ruta}{character. A valid path in your computer where you want to create the csv file. By default the csv named mojo will be created in the working directory.}
}
\value{
It returns a csv in the specified path
}
\description{
This function scraps boxofficemojo.com and downloads box office data of each day in a given interval.
}
