% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movieinfo.R
\name{movieinfo}
\alias{movieinfo}
\title{Scraps mojoboxoffice.com for revenue data}
\usage{
movieinfo(movies, ruta = "~/movieinfo.csv")
}
\arguments{
\item{movies}{character. A string of mojo box office movie ids. You can get them from the id column of a csv extract of mojo() function.}

\item{ruta}{character. A valid path in your computer where you want to create the csv file. By default the csv named movieinfo will be created in the working directory.}
}
\value{
It returns a csv in the specified path
}
\description{
This function scraps mojoboxoffice.com and downloads box office data and other info on a series of movies.
}
